[SG-16] 백엔드 auth api endpoint 구조 변경에 따른 라우트 수정
