+++ core/data/src/main/kotlin/com/captures2024/soongan/core/data/di/RemoteDataSourceModule.kt
@@ -1,7 +1,9 @@
 package com.captures2024.soongan.core.data.di
 
+import com.captures2024.soongan.core.data.remote.AuthDataSource
 import com.captures2024.soongan.core.data.remote.FcmDataSource
 import com.captures2024.soongan.core.data.remote.MembersDataSource
+import com.captures2024.soongan.core.data.remote.impl.AuthDataSourceImpl
 import com.captures2024.soongan.core.data.remote.impl.FcmDataSourceImpl
 import com.captures2024.soongan.core.data.remote.impl.MembersDataSourceImpl
 import dagger.Binds
@@ -21,4 +23,8 @@ abstract class RemoteDataSourceModule {
     @Binds
     @Singleton
     abstract fun bindFcmDataSource(fcmDataSourceImpl: FcmDataSourceImpl): FcmDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindAuthDataSource(authDataSourceImpl: AuthDataSourceImpl): AuthDataSource
 }
