+++ core/data/src/main/kotlin/com/captures2024/soongan/core/data/di/ServiceModule.kt
@@ -1,5 +1,6 @@
 package com.captures2024.soongan.core.data.di
 
+import com.captures2024.soongan.core.data.service.AuthService
 import com.captures2024.soongan.core.data.service.FcmService
 import com.captures2024.soongan.core.data.service.MembersService
 import dagger.Module
@@ -20,4 +21,8 @@ object ServiceModule {
     @Singleton
     @Provides
     internal fun provideFcmService(retrofit: Retrofit): FcmService = retrofit.create(FcmService::class.java)
+
+    @Singleton
+    @Provides
+    internal fun provideAuthService(retrofit: Retrofit): AuthService = retrofit.create(AuthService::class.java)
 }
