+++ core/data/src/main/kotlin/com/captures2024/soongan/core/data/service/MembersService.kt
@@ -18,45 +18,6 @@ import retrofit2.http.Query
 interface MembersService {
 
     /**
-     * 회원 탈퇴 API
-     *
-     * 회원을 탈퇴합니다.
-     **/
-    @Headers("Authorization: true")
-    @POST("members/withdraw")
-    suspend fun withdrawWithToken(): Response<Unit>
-
-    /**
-     * 로그아웃 API
-     *
-     * 로그인시 발급한 JWT를 말소합니다.
-     **/
-    @Headers("Authorization: true")
-    @POST("members/logout")
-    suspend fun signOutWithToken(): Response<Unit>
-
-    /**
-     * 로그인 API
-     *
-     * idToken을 이용하여 로그인을 수행하고, JWT를 발급합니다.
-     **/
-    @POST("members/login")
-    suspend fun signInWithToken(
-        @Body request: SignWithTokenRequest,
-    ): Response<BaseResponse<SignInWithTokenResponse>>
-
-    /**
-     * JWT 갱신 API
-     *
-     * Refresh Token을 이용하여 JWT를 갱신합니다.
-     **/
-    @Headers("Authorization: false")
-    @PATCH("members/refresh")
-    suspend fun reissueToken(
-        @Body request: ReissueTokenRequest,
-    ): Response<BaseResponse<ReissueTokenResponse>>
-
-    /**
      * 프로필 사진 변경 API
      *
      * 프로필 사진을 변경합니다.
