+++ core/domain/src/main/kotlin/com/captures2024/soongan/core/domain/usecase/auth/ReissueTokenUseCase.kt
@@ -1,13 +1,13 @@
-package com.captures2024.soongan.core.domain.usecase.members
+package com.captures2024.soongan.core.domain.usecase.auth
 
-import com.captures2024.soongan.core.data.repository.MembersRepository
+import com.captures2024.soongan.core.data.repository.AuthRepository
 import com.captures2024.soongan.core.domain.runSuspendCatching
 import javax.inject.Inject
 
 class ReissueTokenUseCase
 @Inject
 constructor(
-    private val repository: MembersRepository,
+    private val repository: AuthRepository,
 ) {
 
     suspend operator fun invoke(): Result<Boolean> = runSuspendCatching {
