+++ core/domain/src/main/kotlin/com/captures2024/soongan/core/domain/usecase/auth/SigningGoogleUseCase.kt
@@ -1,6 +1,6 @@
-package com.captures2024.soongan.core.domain.usecase.members
+package com.captures2024.soongan.core.domain.usecase.auth
 
-import com.captures2024.soongan.core.data.repository.MembersRepository
+import com.captures2024.soongan.core.data.repository.AuthRepository
 import com.captures2024.soongan.core.domain.runSuspendCatching
 import com.captures2024.soongan.core.model.network.SocialSignType
 import javax.inject.Inject
@@ -8,7 +8,7 @@ import javax.inject.Inject
 class SigningGoogleUseCase
 @Inject
 constructor(
-    private val repository: MembersRepository,
+    private val repository: AuthRepository,
 ) {
 
     suspend operator fun invoke(
